move code into .c out of .h
